+++ changelog.md
@@ -4,4 +4,7 @@ Beta
 0.1.3 -> 0.1.4
 
 
-- Changes to titan shifting VFX / new particles
+- Real art for odm handles
+- Removed blade sheath recipe, since it's not done
+- Changes to titan shifting VFX / new particles (progress now goes up to 40)
+- Hid the playername and uuid when hovering over a titan's name in chat